Feature v1.0

+++ src/SfCmsProject/CmsBundle/Controller/TemplateController.php
@@ -16,7 +16,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 class TemplateController extends Controller
 {
     /**
-     * @Route("/cms/admin/viewtemplate", name="view_template")
+     * @Route("/cms/admin/security/viewtemplate", name="view_template")
      * @Method({"GET","POST"})
      * @return Response
      */
@@ -31,10 +31,11 @@ class TemplateController extends Controller
             $template = new Template();
             $form = $this->createForm(AddTemplateType::class, $template);
             $formEdit = $this->createForm(EditTemplateType::class, $template);
-
+            $listMedia = $em->getRepository('SfCmsProjectCmsBundle:Image')->findAll();
             $listTemplate = $em->getRepository('SfCmsProjectCmsBundle:Template')->findAll();
             $response = $this->render('SfCmsProjectCmsBundle:Template:viewTemplate.html.twig', array(
                 'listTemplate' => $listTemplate,
+                'listMedia' => $listMedia,
                 'formEdit' => $formEdit->createView(),
                 'form' => $form->createView()))->getContent();
             return new Response($response);
@@ -46,7 +47,7 @@ class TemplateController extends Controller
         }
     }
     /**
-     * @Route("/cms/admin/loadviewtemplate", name="load_view_template")
+     * @Route("/cms/admin/security/loadviewtemplate", name="load_view_template")
      * @Method({"GET","POST"})
      * @return Response
      */
@@ -58,8 +59,14 @@ class TemplateController extends Controller
         if ($request->isXmlHttpRequest()) {
 
             $pageTest = $em->getRepository('SfCmsProjectCmsBundle:Page')->findOneBy(array('name'=>'Titre de la page'));
+            $listPostTest = $em->getRepository('SfCmsProjectCmsBundle:Post')->findBy(array('suppress'=> true));
+            $nbPages = 1;
+            $nbPage = 1;
             $response = $this->render('SfCmsProjectCmsBundle:Template:Custom/'.$request->get('name').'.html.twig', array(
-                'page' => $pageTest
+                'page' => $pageTest,
+                'nbPages' => $nbPages,
+                'nbPage' => $nbPage,
+                'listPost' => $listPostTest
             ))->getContent();
             return new Response($response);
 
@@ -71,7 +78,7 @@ class TemplateController extends Controller
     }
     /**
      * @param Request $request
-     * @Route ("/cms/admin/addtemplatevalid", name="add_template_valid")
+     * @Route ("/cms/admin/security/addtemplatevalid", name="add_template_valid")
      * @Method({"GET","POST"})
      * @return Response
      */
@@ -84,23 +91,27 @@ class TemplateController extends Controller
         // Si la requête est en Ajax
         if ($request->isXmlHttpRequest()) {
 
+            $token = $request->get('csrf');
 
-            $template->setName(str_replace(' ','-',$request->get('name')));
-            $template->setContent($request->get('content'));
+            if ($this->isCsrfTokenValid('csrf_template_page', $token)) {
 
-            $em->persist($template);
-            $em->flush();
+                $template->setName(str_replace(' ', '-', $request->get('name')));
+                $template->setContent($request->get('content'));
 
+                $em->persist($template);
+                $em->flush();
+            }
             // Appel du service pour créer une vue du nouveau template
             $this->container->get('sf_cms_project_cms.CreateTemplateFile')->createTemplateFile($request->get('name'),$request->get('content'));
 
             // Creation d'un nouvel objet template
             $template = new Template();
             $form = $this->createForm(AddTemplateType::class, $template);
-
+            $listMedia = $em->getRepository('SfCmsProjectCmsBundle:Image')->findAll();
             $listTemplate = $em->getRepository('SfCmsProjectCmsBundle:Template')->findAll();
             $response = $this->render('SfCmsProjectCmsBundle:Template:viewTemplate.html.twig', array(
                 'listTemplate' => $listTemplate,
+                'listMedia' => $listMedia,
                 'formEdit' => $formEdit->createView(),
                 'form' => $form->createView()))->getContent();
             return new Response($response);
@@ -114,7 +125,7 @@ class TemplateController extends Controller
     }
     /**
      * @param Request $request
-     * @Route ("/cms/admin/edittemplatevalid", name="edit_template_valid")
+     * @Route ("/cms/admin/security/edittemplatevalid", name="edit_template_valid")
      * @Method({"GET","POST"})
      * @return Response
      */
@@ -124,29 +135,27 @@ class TemplateController extends Controller
         // Si la requête est en Ajax
         if ($request->isXmlHttpRequest()) {
 
-            $this->container->get('sf_cms_project_cms.SupTemplateFile')->supTemplateFile($request->get('presentName'));
-            $template = $em->getRepository('SfCmsProjectCmsBundle:Template')->findOneBy(array('name'=> $request->get('presentName')));
-            $template->setName(str_replace(' ','-',$request->get('name')));
-            $template->setContent($request->get('content'));
+            $token = $request->get('csrf');
 
-            $em->flush();
+            if ($this->isCsrfTokenValid('csrf_edit_template_page', $token)) {
 
-            // Appel du service pour créer une vue twig du nouveau template
-            $this->container->get('sf_cms_project_cms.CreateTemplateFile')->createTemplateFile($request->get('name'),$request->get('content'));
+                $this->container->get('sf_cms_project_cms.SupTemplateFile')->supTemplateFile($request->get('supName'));
+                $template = $em->getRepository('SfCmsProjectCmsBundle:Template')->findOneBy(array('name' => $request->get('supName')));
+                $template->setName(str_replace(' ', '-', $request->get('name')));
+                $template->setContent($request->get('content'));
 
-            //Creation du formulaire d'édition
-            $formEdit = $this->createForm(EditTemplateType::class, $template);
-            // Creation d'un nouvel objet template
-            $template = new Template();
-            $form = $this->createForm(AddTemplateType::class, $template);
+                $em->flush();
 
-            $listTemplate = $em->getRepository('SfCmsProjectCmsBundle:Template')->findAll();
-            $response = $this->render('SfCmsProjectCmsBundle:Template:viewTemplate.html.twig', array(
-                'listTemplate' => $listTemplate,
-                'formEdit' => $formEdit->createView(),
-                'form' => $form->createView()))->getContent();
-            return new Response($response);
 
+                // Appel du service pour créer une vue twig du nouveau template
+                $this->container->get('sf_cms_project_cms.CreateTemplateFile')->createTemplateFile($request->get('name'), $request->get('content'));
+
+                return new Response($template->getName());
+            }
+            else
+            {
+                throw new NotFoundHttpException("La page demandée n'existe pas");
+            }
         }
         else
         {
@@ -156,7 +165,7 @@ class TemplateController extends Controller
     }
     /**
      * @param Request $request
-     * @Route ("/cms/admin/suptemplatevalid", name="sup_template_valid")
+     * @Route ("/cms/admin/security/suptemplatevalid", name="sup_template_valid")
      * @Method({"GET","POST"})
      * @return Response
      */
@@ -166,24 +175,30 @@ class TemplateController extends Controller
         // Si la requête est en Ajax
         if ($request->isXmlHttpRequest()) {
 
-            $this->container->get('sf_cms_project_cms.SupTemplateFile')->supTemplateFile($request->get('name'));
-            $template = $em->getRepository('SfCmsProjectCmsBundle:Template')->findOneBy(array('name'=> $request->get('name')));
-            $listPage = $em->getRepository('SfCmsProjectCmsBundle:Page')->findBy(array('template' => $request->get('name')));
-            foreach ($listPage as $page){
-               $page->setTemplate(null);
-            }
-            $em->remove($template);
-            $em->flush();
+            $token = $request->get('csrf');
 
+            if ($this->isCsrfTokenValid('csrf_sup_template_page', $token)) {
+
+                $this->container->get('sf_cms_project_cms.SupTemplateFile')->supTemplateFile($request->get('name'));
+                $template = $em->getRepository('SfCmsProjectCmsBundle:Template')->findOneBy(array('name' => $request->get('name')));
+                $listPage = $em->getRepository('SfCmsProjectCmsBundle:Page')->findBy(array('template' => $request->get('name')));
+                foreach ($listPage as $page) {
+                    $page->setTemplate(null);
+                }
+                $em->remove($template);
+                $em->flush();
+
+            }
             // Creation d'un nouvel objet template
             $template = new Template();
             $form = $this->createForm(AddTemplateType::class, $template);
             //Creation du formulaire d'édition
             $formEdit = $this->createForm(EditTemplateType::class, $template);
-
+            $listMedia = $em->getRepository('SfCmsProjectCmsBundle:Image')->findAll();
             $listTemplate = $em->getRepository('SfCmsProjectCmsBundle:Template')->findAll();
             $response = $this->render('SfCmsProjectCmsBundle:Template:viewTemplate.html.twig', array(
                 'listTemplate' => $listTemplate,
+                'listMedia' => $listMedia,
                 'formEdit' => $formEdit->createView(),
                 'form' => $form->createView()))->getContent();
             return new Response($response);
@@ -197,7 +212,7 @@ class TemplateController extends Controller
     }
     /**
      * @param Request $request
-     * @Route ("/cms/admin/updateformedit", name="update_form_edit")
+     * @Route ("/cms/admin/security/updateformedit", name="update_form_edit")
      * @Method({"GET","POST"})
      * @return Response
      */
@@ -210,8 +225,10 @@ class TemplateController extends Controller
             $template = $em->getRepository('SfCmsProjectCmsBundle:Template')->findOneBy(array('name'=> $request->get('presentName')));
             //Creation du formulaire d'édition
             $formEdit = $this->createForm(EditTemplateType::class, $template);
+            $listMedia = $em->getRepository('SfCmsProjectCmsBundle:Image')->findAll();
 
             $response = $this->render('SfCmsProjectCmsBundle:Template:formEditTemplatePage.html.twig', array(
+                'listMedia' => $listMedia,
                 'formEdit' => $formEdit->createView()))->getContent();
             return new Response($response);
 
@@ -223,5 +240,29 @@ class TemplateController extends Controller
 
     }
 
+    /**
+     * @param Request $request
+     * @Route ("/cms/admin/security/reloadlisttemplate", name="reload_list_template")
+     * @Method({"GET","POST"})
+     * @return Response
+     */
+    public function reloadListTemplateAction(Request $request){
+
+        $em = $this->getDoctrine()->getManager();
+        // Si la requête est en Ajax
+        if ($request->isXmlHttpRequest()) {
+
+            $listTemplate = $em->getRepository('SfCmsProjectCmsBundle:Template')->findAll();
+            $response = $this->render('SfCmsProjectCmsBundle:Template:listTemplatePage.html.twig', array(
+                'listTemplate' => $listTemplate))->getContent();
+            return new Response($response);
+
+        }
+        else
+        {
+            throw new NotFoundHttpException("La page demandée n'existe pas");
+        }
+
+    }
 
 }
